Conversion warning fixes

+++ main.cpp
@@ -60,13 +60,13 @@ using ll = long long;
 
 constexpr int dx[] = {-1, 0, 0, 1, -1, -1, 1, 1};
 constexpr int dy[] = {0, -1, 1, 0, -1, 1, -1, 1};
-constexpr int INF = 1e9 + 15;
-constexpr ll INFL = 1e18 + 15;
+constexpr int INF = (int)1e9 + 15;
+constexpr ll INFL = ll(1e18) + 15;
 constexpr double EPS = 1e-9;
 constexpr int P = 31;
 constexpr int P2 = 33;
-constexpr int M = 1e9 + 7;
-constexpr int M2 = 1e9 + 9;
+constexpr int M = (int)1e9 + 7;
+constexpr int M2 = (int)1e9 + 9;
 
 constexpr int gcd(int a, int b) {
   if (b == 0)
